fix: Minor changes

+++ bank-server/src/main/java/com/leebuntu/server/handler/admin/banking/AccountHandler.java
@@ -11,6 +11,8 @@ import com.leebuntu.common.banking.dto.request.banking.RemoveAccount;
 import com.leebuntu.common.banking.dto.request.banking.ViewAccount;
 import com.leebuntu.common.banking.dto.response.banking.Accounts;
 import com.leebuntu.server.communication.router.ContextHandler;
+import com.leebuntu.server.db.core.Database;
+import com.leebuntu.server.db.core.DatabaseManager;
 import com.leebuntu.server.provider.AccountProvider;
 import com.leebuntu.server.provider.CustomerProvider;
 
@@ -18,6 +20,7 @@ import java.time.Instant;
 import java.util.List;
 
 public class AccountHandler {
+    private static final Database accountDB = DatabaseManager.getDB("accounts");
 
     private static BankingResult createCheckingAccount(Account account) {
         if (AccountProvider.isAccountNumberExist(account.getAccountNumber())) {
@@ -54,32 +57,26 @@ public class AccountHandler {
 
     public static ContextHandler createAccount() {
         return (context) -> {
+            if (!CustomerProvider.isAdmin((int) context.getField("userId"))) {
+                context.reply(new Response(Status.NOT_AUTHORIZED, "Unauthorized"));
+                return;
+            }
+
             CreateAccount request = new CreateAccount();
 
             if (context.bind(request)) {
+                Account account = request.getAccount();
+                if (!account.validate()) {
+                    context.reply(new Response(Status.FAILED, "Invalid account data"));
+                    return;
+                }
+
                 BankingResult result = null;
 
                 try {
-                    if (request.getAccount().getAccountType() == AccountType.CHECKING) {
-                        Account account = new Account();
-                        account.setCustomerId(request.getAccount().getCustomerId());
-                        account.setAccountNumber(request.getAccount().getAccountNumber());
-                        account.setTotalBalance(request.getAccount().getTotalBalance());
-                        account.setAvailableBalance(request.getAccount().getAvailableBalance());
-                        account.setOpenDate(Instant.now().toEpochMilli());
-                        account.setAccountType(request.getAccount().getAccountType());
-                        account.setLinkedSavingsAccountNumber(request.getAccount().getLinkedSavingsAccountNumber());
+                    if (account.getAccountType() == AccountType.CHECKING) {
                         result = createCheckingAccount(account);
-                    } else if (request.getAccount().getAccountType() == AccountType.SAVINGS) {
-                        Account account = new Account();
-                        account.setCustomerId(request.getAccount().getCustomerId());
-                        account.setAccountNumber(request.getAccount().getAccountNumber());
-                        account.setTotalBalance(request.getAccount().getTotalBalance());
-                        account.setAvailableBalance(request.getAccount().getAvailableBalance());
-                        account.setOpenDate(Instant.now().toEpochMilli());
-                        account.setAccountType(request.getAccount().getAccountType());
-                        account.setInterestRate(request.getAccount().getInterestRate());
-                        account.setMaxTransferAmountToChecking(request.getAccount().getMaxTransferAmountToChecking());
+                    } else if (account.getAccountType() == AccountType.SAVINGS) {
                         result = createSavingsAccount(account);
                     }
 
@@ -102,6 +99,10 @@ public class AccountHandler {
 
     public static ContextHandler getAllAccounts() {
         return (context) -> {
+            if (!CustomerProvider.isAdmin((int) context.getField("userId"))) {
+                context.reply(new Response(Status.NOT_AUTHORIZED, "Unauthorized"));
+                return;
+            }
             List<Account> accounts = AccountProvider.getAllAccounts();
             if (accounts.isEmpty()) {
                 context.reply(new Response(Status.FAILED, "No accounts found"));
@@ -114,6 +115,11 @@ public class AccountHandler {
 
     public static ContextHandler getAccount() {
         return (context) -> {
+            if (!CustomerProvider.isAdmin((int) context.getField("userId"))) {
+                context.reply(new Response(Status.NOT_AUTHORIZED, "Unauthorized"));
+                return;
+            }
+
             ViewAccount request = new ViewAccount();
             if (context.bind(request)) {
                 int customerId = request.getCustomerId();
@@ -132,13 +138,93 @@ public class AccountHandler {
 
     public static ContextHandler deleteAccount() {
         return (context) -> {
+            if (!CustomerProvider.isAdmin((int) context.getField("userId"))) {
+                context.reply(new Response(Status.NOT_AUTHORIZED, "Unauthorized"));
+                return;
+            }
+
             RemoveAccount request = new RemoveAccount();
             if (context.bind(request)) {
-                if (AccountProvider.deleteAccount(request.getAccountNumber())) {
-                    context.reply(new Response(Status.SUCCESS, "Account deleted successfully"));
-                } else {
-                    context.reply(new Response(Status.FAILED, "Failed to delete account"));
+                AccountType accountType = AccountProvider.getAccountType(request.getAccountNumber());
+
+                switch (accountType) {
+                    case CHECKING:
+                        if (AccountProvider.deleteAccount(request.getAccountNumber())) {
+                            context.reply(new Response(Status.SUCCESS, "Account deleted successfully"));
+                        } else {
+                            context.reply(new Response(Status.FAILED, "Failed to delete account"));
+                        }
+                        break;
+                    case SAVINGS:
+                        accountDB.beginTransaction();
+                        List<Account> linkedCheckingAccounts = AccountProvider
+                                .getAccountByLinkedSavingsAccountNumber(request.getAccountNumber());
+                        for (Account linkedCheckingAccount : linkedCheckingAccounts) {
+                            linkedCheckingAccount.setLinkedSavingsAccountNumber(null);
+                            AccountProvider.updateAccount(linkedCheckingAccount);
+                        }
+
+                        if (AccountProvider.deleteAccount(request.getAccountNumber())) {
+                            accountDB.endTransaction();
+                            context.reply(new Response(Status.SUCCESS, "Account deleted successfully"));
+                        } else {
+                            accountDB.endTransaction();
+                            context.reply(new Response(Status.FAILED, "Failed to delete account"));
+                        }
+
+                        break;
                 }
+
+            } else {
+                context.reply(new Response(Status.FAILED, "Failed to bind request"));
+            }
+        };
+    }
+
+    public static ContextHandler updateAccount() {
+        return (context) -> {
+            if (!CustomerProvider.isAdmin((int) context.getField("userId"))) {
+                context.reply(new Response(Status.NOT_AUTHORIZED, "Unauthorized"));
+                return;
+            }
+
+            CreateAccount request = new CreateAccount();
+
+            if (context.bind(request)) {
+                Account account = request.getAccount();
+                if (account.getAccountNumber() != null && !account.getAccountNumber().isEmpty()) {
+                    account.setAccountType(AccountProvider.getAccountType(account.getAccountNumber()));
+                }
+
+                if (!account.validate()) {
+                    context.reply(new Response(Status.FAILED, "Invalid account data"));
+                    return;
+                }
+
+                if (account.getAccountType() == AccountType.CHECKING) {
+                    if (!account.getLinkedSavingsAccountNumber().isEmpty()) {
+                        if (!AccountProvider.isAccountNumberExist(account.getLinkedSavingsAccountNumber())
+                                || !AccountProvider.isAccountOwner(account.getLinkedSavingsAccountNumber(),
+                                        account.getId())) {
+                            context.reply(new Response(Status.FAILED, "Savings account does not exist"));
+                            return;
+                        }
+                    }
+                }
+
+                try {
+                    boolean result = AccountProvider.updateAccount(account);
+
+                    if (result) {
+                        context.reply(new Response(Status.SUCCESS, "Account updated successfully"));
+                    } else {
+                        context.reply(new Response(Status.FAILED, "Failed to update account"));
+                    }
+                } catch (Exception e) {
+                    context.reply(new Response(Status.FAILED, "Failed to bind request"));
+                    return;
+                }
+
             } else {
                 context.reply(new Response(Status.FAILED, "Failed to bind request"));
             }
